added env file and made helper function file
