+++ server/reviewsController.js
@@ -4,6 +4,7 @@
 /* eslint-disable no-plusplus */
 /* eslint-disable no-console */
 const { db } = require('../database/index');
+const helpers = require('./reviewsHelpers');
 
 const getReviews = (req, res) => {
   const { productId } = req.params;
@@ -45,96 +46,7 @@ const getReviews = (req, res) => {
         }
       }
 
-      const extractPhotos = (reviews) => {
-        const photos = [];
-        const tracker = [];
-
-        for (let i = 0; i < reviews.length; i++) {
-          const item = reviews[i];
-
-          const obj = {
-            id: item.photo_id,
-            url: item.photo_url,
-          };
-
-          if (!tracker.includes(item.photo_url) && item.photo_url) {
-            photos.push(obj);
-          }
-
-          tracker.push(item.photo_url);
-        }
-
-        return photos;
-      };
-
-      const extractCharacteristics = (reviews) => {
-        const chars = {};
-
-        for (let i = 0; i < reviews.length; i++) {
-          const item = reviews[i];
-
-          if (!chars[item.characteristics_name]) {
-            chars[item.characteristics_name] = [
-              item.characteristics_id,
-              item.characteristics_value,
-            ];
-          } else if (chars[item.characteristics_name]) {
-            chars[item.characteristics_name].push(item.characteristics_value);
-          }
-        }
-
-        const getAverage = (arr) => arr.reduce((acc, n) => acc + n) / arr.length;
-
-        for (const key in chars) {
-          const item = chars[key];
-          const obj = { id: 0, value: '' };
-
-          [obj.id] = item;
-          obj.value = getAverage(item.slice(1)).toString();
-
-          chars[key] = obj;
-        }
-
-        return chars;
-      };
-
-      const constructMeta = (reviews, tracker) => {
-        const reviewsUniq = [];
-
-        for (const key in tracker) {
-          reviewsUniq.push(tracker[key][0]);
-        }
-
-        const meta = {
-          product_id: productId.toString(),
-          ratings: {},
-          recommended: {
-            true: 0,
-            false: 0,
-          },
-          characteristics: extractCharacteristics(reviews),
-        };
-
-        for (let i = 0; i < reviewsUniq.length; i++) {
-          const item = reviewsUniq[i];
-
-          if (!meta.ratings[item.rating]) {
-            meta.ratings[item.rating] = 1;
-          } else if (meta.ratings[item.rating]) {
-            meta.ratings[item.rating]++;
-          }
-
-          for (const key in meta.ratings) {
-            meta.ratings[key] = meta.ratings[key].toString();
-          }
-
-          meta.recommended[item.recommend]++;
-        }
-
-        return meta;
-      };
-
-      output.meta = constructMeta(result, reviewsTracker);
+      output.meta = helpers.constructMeta(result, reviewsTracker, productId);
 
       for (const review in reviewsTracker) {
         const item = reviewsTracker[review][0];
@@ -154,7 +66,7 @@ const getReviews = (req, res) => {
             date: JSON.stringify(item.date).slice(1, -1),
             reviewer_name: item.reviewer_name,
             helpfulness: item.helpfulness,
-            photos: extractPhotos(reviewsTracker[review]),
+            photos: helpers.extractPhotos(reviewsTracker[review]),
           };
 
           output.results.push(reviewObj);
@@ -206,33 +118,6 @@ const reportReview = (req, res) => {
 const postNewReview = (req, res) => {
   const data = req.body;
 
-  const constructPhotoQueries = (photos) => {
-    let output = 'VALUES ';
-
-    for (let i = 0; i < photos.length; i++) {
-      const count = i + 1;
-      const str = `((SELECT max(id) FROM reviews_photos) + ${count}, (SELECT id FROM ins1), '${photos[i]}'),`;
-
-      output += str;
-    }
-
-    return output.slice(0, -1);
-  };
-
-  const constructCharacteristicQueries = (chars) => {
-    let output = 'VALUES ';
-    let count = 1;
-
-    for (const key in chars) {
-      const str = `((SELECT max(id) FROM reviews_characteristics) + ${count}, ${key}, (SELECT id FROM ins1), ${chars[key]}),`;
-
-      output += str;
-      count++;
-    }
-
-    return output.slice(0, -1);
-  };
-
   const reviewQuery = `
   WITH ins1 AS (
     INSERT INTO reviews(id, product_id, rating, date, summary, body, recommend, reported, reviewer_name, reviewer_email, response, helpfulness)
@@ -241,10 +126,10 @@ const postNewReview = (req, res) => {
     )
   ,ins2 AS (
     INSERT INTO reviews_photos(id, review_id, url)
-    ${constructPhotoQueries(data.photos)}
+    ${helpers.constructPhotoQueries(data.photos)}
     )
   INSERT INTO reviews_characteristics(id, characteristic_id, review_id, value)
-  ${constructCharacteristicQueries(data.characteristics)}
+  ${helpers.constructCharacteristicQueries(data.characteristics)}
   RETURNING (SELECT id AS review_id FROM ins1)
   ;`;
 
@@ -266,6 +151,8 @@ module.exports = {
 
 /*
 TODO:
+-break db queries out into db file, export them as functions
+-allow user to configure db connection using env file (currently hardcoded)
 -add evergreen data for potential failing requests
 -add tests (K6?)
 -implement 'sort by' feature (relevance, newness, helpfulness)
